==> RNAseq_Analysis/Differential_Editing/JACUSA2/3_Split_samples.ipynb <==
This notebook simply splits the JACUSA2 output by sample for ease of manipulation before further filtering occurs.

data <- read.table('/mnt/vast/hpc/csg/hcs2152/ZFR_RNA_Editing/JACUSA2/all_dpf/JACUSA2_all_dpf_sites_no_SNPS.txt',header=TRUE,sep='\t')

# Creating separate data frames for each 'bases' column and renaming them
bases11_df <- data[, c("ID", "strand", "ref", "bases11")]
colnames(bases11_df) <- c("ID", "strand", "ref", "Base_counts")

bases12_df <- data[, c("ID", "strand", "ref", "bases12")]
colnames(bases12_df) <- c("ID", "strand", "ref", "Base_counts")

bases13_df <- data[, c("ID", "strand", "ref", "bases13")]
colnames(bases13_df) <- c("ID", "strand", "ref", "Base_counts")

bases14_df <- data[, c("ID", "strand", "ref", "bases14")]
colnames(bases14_df) <- c("ID", "strand", "ref", "Base_counts")

bases15_df <- data[, c("ID", "strand", "ref", "bases15")]
colnames(bases15_df) <- c("ID", "strand", "ref", "Base_counts")

bases16_df <- data[, c("ID", "strand", "ref", "bases16")]
colnames(bases16_df) <- c("ID", "strand", "ref", "Base_counts")

bases21_df <- data[, c("ID", "strand", "ref", "bases21")]
colnames(bases21_df) <- c("ID","strand", "ref", "Base_counts")

bases22_df <- data[, c("ID","strand", "ref", "bases22")]
colnames(bases22_df) <- c("ID","strand", "ref", "Base_counts")

bases23_df <- data[, c("ID","strand", "ref", "bases23")]
colnames(bases23_df) <- c("ID","strand", "ref", "Base_counts")

bases24_df <- data[, c("ID","strand", "ref", "bases24")]
colnames(bases24_df) <- c("ID","strand", "ref", "Base_counts")

bases25_df <- data[, c("ID","strand", "ref", "bases25")]
colnames(bases25_df) <- c("ID","strand", "ref", "Base_counts")

bases26_df <- data[, c("ID","strand", "ref", "bases26")]
colnames(bases26_df) <- c("ID","strand", "ref", "Base_counts")

output_directory <- '/mnt/vast/hpc/csg/hcs2152/ZFR_RNA_Editing/JACUSA2/all_dpf/all_dpf_singles'

write.table(bases11_df, file.path(output_directory, 'Ctrl-01_bases.tsv'), row.names = FALSE, sep = '\t')
write.table(bases12_df, file.path(output_directory, 'Ctrl-02_bases.tsv'), row.names = FALSE, sep = '\t')
write.table(bases13_df, file.path(output_directory, 'Ctrl-03_bases.tsv'), row.names = FALSE, sep = '\t')
write.table(bases14_df, file.path(output_directory, 'Ctrl-04_bases.tsv'), row.names = FALSE, sep = '\t')
write.table(bases15_df, file.path(output_directory, 'Ctrl-05_bases.tsv'), row.names = FALSE, sep = '\t')
write.table(bases16_df, file.path(output_directory, 'Ctrl-06_bases.tsv'), row.names = FALSE, sep = '\t')
write.table(bases21_df, file.path(output_directory, 'NO-01_bases.tsv'), row.names = FALSE, sep = '\t')
write.table(bases22_df, file.path(output_directory, 'NO-02_bases.tsv'), row.names = FALSE, sep = '\t')
write.table(bases23_df, file.path(output_directory, 'NO-03_bases.tsv'), row.names = FALSE, sep = '\t')
write.table(bases24_df, file.path(output_directory, 'NO-04_bases.tsv'), row.names = FALSE, sep = '\t')
write.table(bases25_df, file.path(output_directory, 'NO-05_bases.tsv'), row.names = FALSE, sep = '\t')
write.table(bases26_df, file.path(output_directory, 'NO-06_bases.tsv'), row.names = FALSE, sep = '\t')
