==> RNAseq_Analysis/Differential_Editing/JACUSA2/4_Calculate_proportions.ipynb <==
### Filtering JACUSA2 Output
This script calculates editing proprtions of each site per sample such that further filtering can be done.


# Load required libraries
library(dplyr)
library(readr)

process_file <- function(file_path, output_directory) {
  # Read the TSV file
  df <- read.table(file_path, header = TRUE, sep = "\t", stringsAsFactors = FALSE)

  # Split the base counts column
  counts <- strsplit(df$Base_counts, ",")
  
  # Convert the list to a data frame
  counts_df <- as.data.frame(do.call(rbind, counts))
  
  # Rename the columns
  colnames(counts_df) <- c("A_count", "C_count", "G_count", "T_count")
  
  # Combine the new columns with the original data frame
  df <- cbind(df, counts_df)
  
  
  # Remove unnecessary columns
  df <- df[, !colnames(df) %in% c("Base_counts")]
  
  # Convert counts columns to numeric
  df[, c("A_count", "C_count", "G_count", "T_count")] <- lapply(
    df[, c("A_count", "C_count", "G_count", "T_count")], as.numeric)
  
  
  # Calculate Non_Edited_Count and Edited_Count
  edited_df <- df %>%
    mutate(
      Non_Edited_Count = case_when(
        ref == "A" ~ A_count,
        ref == "C" ~ C_count,
        ref == "G" ~ G_count,
        ref == "T" ~ T_count
      ),
      Edited_Count = rowSums(select(., A_count, C_count, G_count, T_count), na.rm = TRUE) - as.numeric(Non_Edited_Count)
    ) %>%
    select(ID, Non_Edited_Count, Edited_Count)
  
  # Calculate Edited_Count_Proportion
  edited_df$Edited_Count_Proportion <- edited_df$Edited_Count / (edited_df$Edited_Count + edited_df$Non_Edited_Count)
  
  print("Final DataFrame:")
  print(head(edited_df))
  
  # Get the base name of the original file
  base_name <- tools::file_path_sans_ext(basename(file_path))
  
  # Create the output file path
  output_file_path <- file.path(output_directory, paste0(base_name, "___Edited_Column_Added.tsv"))
  
  # Write the processed table to a TSV file
  write.table(edited_df, file = output_file_path, sep = "\t", quote = FALSE, row.names = FALSE)
}

input_directory_path <- '/mnt/vast/hpc/csg/hcs2152/ZFR_RNA_Editing/JACUSA2/all_dpf/all_dpf_singles'
output_directory_path <- '/mnt/vast/hpc/csg/hcs2152/ZFR_RNA_Editing/JACUSA2/all_dpf/all_dpf_singles/Edited'

# Get a list of files in the input directory
file_list <- list.files(input_directory_path, pattern = "\\.tsv$", full.names = TRUE)
print(file_list)

for (file in file_list) {
  process_file(file, output_directory_path)
}
