==> RNAseq_Analysis/Differential_Editing/JACUSA2/5_Rename_columns.ipynb <==
This notebook renames all of the Editing/Non-editing and proprtion samples so we can merge the samples into one dataframe and be able to distinguish by sample.

library(tidyverse)

input_directory_path <- '/mnt/vast/hpc/csg/hcs2152/ZFR_RNA_Editing/JACUSA2/all_dpf/all_dpf_singles/Edited'
output_directory_path <- '/mnt/vast/hpc/csg/hcs2152/ZFR_RNA_Editing/JACUSA2/all_dpf/all_dpf_singles/Edited'

# Get a list of files in the input directory
file_list <- list.files(input_directory_path, pattern = "\\___Edited_Column_Added.tsv$", full.names = TRUE)

# Function to process each file
process_file <- function(file_path, output_directory) {
  # Print the file path being processed
  cat("Processing file:", file_path, "\n")
  
  # Read the TSV file
  df <- read.table(file_path, header = TRUE, sep = "\t", stringsAsFactors = FALSE)
  
  # Extract base name before the first underscore
  base_name <- tools::file_path_sans_ext(basename(file_path))
  
  cat("Head of the DataFrame for file:", base_name, "\n")
  print(head(df))
  
  # Extract the prefix before the second underscore
  parts <- unlist(strsplit(base_name, "_"))
  prefix <- paste(parts[1:2], collapse = "_")
  
  # Print the extracted prefix
  cat("Extracted Prefix for file:", base_name, "is", prefix, "\n")
  
  # Add the base name to specified column names
  colnames(df)[2:4] <- paste0(prefix, "_", colnames(df)[2:4])

  # Print the processed data frame
  cat("Processed DataFrame for file:", base_name, "\n")
  print(head(df))
  
  # Write the processed table to a new TSV file in the subfolder
  output_file_path <- file.path(output_directory_path, paste0(base_name, "___Renamed.tsv"))
  write.table(df, file = output_file_path, sep = "\t", quote = FALSE, row.names = FALSE)
}

# Process each file in the directory
for (file_path in file_list) {
  process_file(file_path)
}
