==> RNAseq_Analysis/Differential_Editing/JACUSA2/6_Merge_data.ipynb <==
This notebook merges samples from a single DPF.

library(dplyr)
library(purrr)
library(tidyverse)

input_directory <- '/mnt/vast/hpc/csg/hcs2152/ZFR_RNA_Editing/JACUSA2/all_dpf/all_dpf_singles/Edited'

# List all files in the folder
all_files <- list.files(input_directory, full.names = TRUE)

#Select only renamed files
pattern <- "___Edited_Column_Added___Renamed.tsv$"
# Use grep to filter files based on the pattern
selected_files <- grep(pattern, all_files, value = TRUE)

print(selected_files)

# Initialize an empty list to store data frames
df_list <- list()

# Read each file and store in the list
for (file_path in selected_files) {
  df <- read.table(file_path, header = TRUE, sep = "\t", stringsAsFactors = FALSE)
  df_list[[basename(file_path)]] <- df
}

# Merge data frames by the intersecting "ID" column
merged_df <- reduce(df_list, inner_join, by = "ID")

merged_df <- data.frame(merged_df, col.names = TRUE)

write.table(merged_df, "/mnt/vast/hpc/csg/hcs2152/ZFR_RNA_Editing/JACUSA2/all_dpf/all_dpf_merged_data.tsv", sep = "\t", row.names = FALSE, col.names = TRUE, quote = FALSE)
