==> RNAseq_Analysis/Differential_Editing/JACUSA2/10_Prep_REDITs_input_2.ipynb <==
This is a continuation of Step 9 to format the REDITs input. This script pivots the Edited/Non-edited counts per sample.

library(tidyr)
library(dplyr)
library(readr)
library(stringr)

#Specify input directories
AG_path <- "/mnt/vast/hpc/csg/hcs2152/ZFR_RNA_Editing/JACUSA2/REDITs/all_dpf/AG"
TC_path <- "/mnt/vast/hpc/csg/hcs2152/ZFR_RNA_Editing/JACUSA2/REDITs/all_dpf/TC"

#Specify output directories
AG_output_dir <- "/mnt/vast/hpc/csg/hcs2152/ZFR_RNA_Editing/JACUSA2/REDITs/all_dpf/AG/Pivoted"
TC_output_dir <- "/mnt/vast/hpc/csg/hcs2152/ZFR_RNA_Editing/JACUSA2/REDITs/all_dpf/TC/Pivoted"

process_file <- function(file_path, output_directory) {
  # Extract prefix from the file name
  prefix <- sub("^(.*?)_", "\\1", tools::file_path_sans_ext(basename(file_path)))
  
  # Read the TSV file
  df <- read.table(file_path, header = TRUE, sep = "\t", stringsAsFactors = FALSE)
  df <- df %>%
    select(1, 3, 2, everything())

  cat("Head of the input file", prefix, ":\n")
  print(head(df))

  # Pivot and manipulate the data
  stacked_df <- df %>%
    pivot_longer(
        cols = -ID,  # Specify the columns to pivot
        names_to = "Count_Type",
        values_to = "Counts"
    ) %>%
    mutate(
        Count_Type = rep(c("Edited", "Non_Edited"), length.out = n()),
        ID = factor(ID, levels = unique(ID))
    ) %>%
    arrange(ID, Count_Type)

  
  # Rename the "Counts" column to "Prefix_Counts"
  colnames(stacked_df)[colnames(stacked_df) == "Counts"] <- paste(prefix, "Counts", sep = "_")
  
  # Create the final stacked data frame
  stacked_df <- stacked_df %>%
    mutate(ID___Count_Type = paste(ID, Count_Type, sep = "___")) %>%
    select(ID___Count_Type, everything()) %>%
    select(-c(ID, Count_Type))

  cat("Head of the stacked data frame final", prefix, ":\n")
  print(head(stacked_df))
  
  # Write the result to a TSV file in the output directory
  output_file_path <- file.path(output_directory, paste(prefix, "stacked.tsv", sep = "_"))
  write.table(stacked_df, file = output_file_path, sep = "\t", quote = FALSE, row.names = FALSE)
  # Print a success message
  cat("*Table written successfully to:*", output_file_path, "\n")
  
}

# List all TSV files in the directory
file_list_AG <- list.files(AG_path, pattern = ".tsv$", full.names = TRUE)
file_list_TC <- list.files(TC_path, pattern = ".tsv$", full.names = TRUE)

print(file_list_AG)
print(file_list_TC)

# Process each file and store the results in the output directory
lapply(file_list_AG, function(file_path) {
  process_file(file_path, AG_output_dir)
})

lapply(file_list_TC, function(file_path) {
  process_file(file_path, TC_output_dir)
})
