==> RNAseq_Analysis/Differential_Editing/JACUSA2/11_Prep_REDITs_input_3.ipynb <==
This script merges all of the samples into one for REDITs analysis. 

library(dplyr)
library(purrr)
library(tidyverse)

# Specify the input directory containing TSV files that were pivoted
AG_input <- "/mnt/vast/hpc/csg/hcs2152/ZFR_RNA_Editing/JACUSA2/REDITs/all_dpf/AG/Pivoted"
TC_input <- "/mnt/vast/hpc/csg/hcs2152/ZFR_RNA_Editing/JACUSA2/REDITs/all_dpf/TC/Pivoted"

Since we are merging A/G edited and T/C edited sites, we are going to add a tag onto the ID for each before the merge to keep track.

AG_files <- list.files(AG_input, pattern = "\\.tsv$", full.names = TRUE)
TC_files <- list.files(TC_input, pattern = "\\.tsv$", full.names = TRUE)

# Add tag to AG files
for (file_path in AG_files) {
  # Read the TSV file into a data frame
  df <- read_tsv(file_path)

  # Add "_AG" at the end of the "ID___Count_Type" column values
  df <- df %>%
    mutate(ID___Count_Type = paste(ID___Count_Type, "_AG", sep = ""))

  # Save the modified data frame back to the file
  write_tsv(df, file_path)
}

#Add tag to TC files
for (file_path in TC_files) {
  # Read the TSV file into a data frame
  df <- read_tsv(file_path)

  # Add "_TC" at the end of the "ID___Count_Type" column values
  df <- df %>%
    mutate(ID___Count_Type = paste(ID___Count_Type, "_TC", sep = ""))

  # Save the modified data frame back to the file
  write_tsv(df, file_path)
}

#Define an output directory
output_directory <- "/mnt/vast/hpc/csg/hcs2152/ZFR_RNA_Editing/JACUSA2/REDITs/all_dpf/Merged"

# Convert to data list
AG_data_list <- lapply(AG_files, function(file) read.delim(file, sep = "\t"))
TC_data_list <- lapply(TC_files, function(file) read.delim(file, sep = "\t"))

# Merge by intersect only for AG and TC separately first.
AG_merged_data <- Reduce(function(x, y) merge(x, y, by = "ID___Count_Type", all = FALSE), AG_data_list)
TC_merged_data <- Reduce(function(x, y) merge(x, y, by = "ID___Count_Type", all = FALSE), TC_data_list)

# Now merge AG and TC dataframes 
combined_df <- bind_rows(AG_merged_data, TC_merged_data)

combined_df

# Write merged_data to a TSV file
write.table(combined_df, file = file.path(output_directory, "JACUSA2_all_dpf_REDITs_input.tsv"), sep = "\t", quote = FALSE, row.names = FALSE)
