==> RNAseq_Analysis/Differential_Editing/JACUSA2/12_REDITs_LLR.ipynb <==
library(doParallel)
library(foreach)
library(stringr)
source("/home/hcs2152/github/REDITs/REDIT_LLR.R")

# Input data 
file_path <- '/mnt/vast/hpc/csg/hcs2152/ZFR_RNA_Editing/JACUSA2/REDITs/5dpf/Merged/JACUSA2_5dpf_REDITs_input.tsv'

# Read the TSV file into a data frame
data <- read.table(file_path, header = TRUE, sep = "\t")

#initiate a cluster
noCores = detectCores() -1
cl = makeCluster(noCores,outfile="")
registerDoParallel(cl,cores=noCores)

data

# Extract the first column (index column)
ids <- data[, 1]

# Exclude the first column for further processing
data <- data[, -1]

data

#Split up the table into separate matrices for processing
rows_per_matrix <- 2

matrix_list <- split(data, (seq(nrow(data))-1) %/% rows_per_matrix)

#Modify the ID column to append after REDITs has ran on data
# Extract the desired part of the column
new_ids <- data.frame(ID = str_extract(ids, "(\\d+_\\d+)___\\w+_(AG|TC)"), stringsAsFactors = FALSE)

# Remove the middle part of the 'ID' values
new_ids$ID <- gsub("_(Edited|Non_Edited)_", "_", new_ids$ID)

# Remove duplicates based on the modified 'ID' column
new_ids <- unique(new_ids)

new_ids

# Assuming matrix_list is a list of data frames
matrix_list <- lapply(matrix_list, as.matrix)

# Check the number of matrices in matrix_list
num_editing_sites <- length(matrix_list)

# Print the result
print(num_editing_sites)

#Set the groups for each matrix 
groups=c('ctrl','ctrl','ctrl','mutant','mutant','mutant')

output_matrix = foreach(i=1:num_editing_sites,.combine='rbind') %dopar%{
    current_matrix <- matrix_list[[i]]
    regression_info = REDIT_LLR(data=current_matrix, groups=groups)
    return( as.matrix( data.frame(p_value= regression_info$p.value )) )
}

# Combine IDs with p-values
p_vals_per_site <- cbind(new_ids,output_matrix)

# Check
p_vals_per_site

# Adjust p_values
p_vals_per_site$p_adj_BH <- p.adjust(p_vals_per_site$p_value, method = "BH")

p_vals_per_site

p_file <- "/mnt/vast/hpc/csg/hcs2152/ZFR_RNA_Editing/JACUSA2/REDITs/5dpf/JACUSA2_5dpf_p_values.tsv"
write.table(p_vals_per_site, file = p_file, sep = "\t", quote = FALSE, row.names = FALSE)
