==> RNAseq_Analysis/Differential_Editing/JACUSA2/17_Annotate_VEP_p_values.ipynb <==
This notebook merges the VEP "Most Severe" with the associated p-values from REDITs.

library(dplyr)
library(tidyr)

# Set paths to data
vep_file <- "/mnt/vast/hpc/csg/hcs2152/ZFR_RNA_Editing/JACUSA2/VEP/all_dpf/JACUSA2_all_dpf_VEP_most_severe_output.txt"
p_file <- "/mnt/vast/hpc/csg/hcs2152/ZFR_RNA_Editing/JACUSA2/REDITs/all_dpf/JACUSA2_all_dpf_p_values.tsv"

# Load in data
vep_data <- read.table(vep_file, header = FALSE, sep = "\t", stringsAsFactors = FALSE)
p_data <- read.table(p_file, header = TRUE, sep = "\t", stringsAsFactors = FALSE)

#Extract ID and most severe consequence columns and name accordingly 
vep_new <- vep_data[, c(1, 4)]
colnames(vep_new) <- c("ID", "most_severe_consequence")

#Remove the _A/G from ID column
vep_new$ID <- sub("_A/G", "", vep_new$ID)

p_data$ID <- sub("___AG", "", p_data$ID)
p_data$ID <- sub("___TC", "", p_data$ID)

annotated_data <- merge(p_data, vep_new, by = "ID")

# Remove p_adj values (redoing after filter)
annotated_data <- annotated_data[, !colnames(annotated_data) %in% c("p_adj_BH")]

annotated_data

output_file <- "/mnt/vast/hpc/csg/hcs2152/ZFR_RNA_Editing/JACUSA2/all_dpf/JACUSA2_all_dpf_VEP_p_val.tsv"

write.table(annotated_data, file = output_file, sep = "\t", quote = FALSE, row.names = FALSE)
